Refactor game files structure
